requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0
PyGithub>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.0.0
schedule>=1.2.0 